papers/fock_state_expressivity/requirements.txt
numpy
scikit-learn
matplotlib
merlinquantum
pytest
torch
tqdm